Coreライブラリの作成と関数・クラスの移行
