+++ ytdlpSpout/__init__.py
@@ -1,0 +1,22 @@
+
+from .core import (
+    Streamer,
+    get_executable_dir,
+    find_ffmpeg_path,
+    check_av1_support,
+    get_optimal_format_string,
+    build_ffmpeg_header_args,
+    detect_fps,
+    detect_max_resolution,
+)
+
+__all__ = [
+    "Streamer",
+    "get_executable_dir",
+    "find_ffmpeg_path",
+    "check_av1_support",
+    "get_optimal_format_string",
+    "build_ffmpeg_header_args",
+    "detect_fps",
+    "detect_max_resolution",
+]
